Refactor install-apache-and-php.yml
Consolidate the playbook by using the package module for generic package manager and
used variables for the package which will install on each remote machine.
The variable are defined in the inventory. Each remote host has it own variable.
This elimates the need for checking if the ansible_distribution is Ubuntu or CentOS